Organized components and Category almost done
