+++ src/components/MainPublicity/MainPublicity.css
