+++ src/components/RestaurantTitle/RestaurantTitle.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import './styles/RestaurantTitle.css';
-import hamburgerIcon from '../assets/icons/1046784.svg';
+import './RestaurantTitle.css';
+import hamburgerIcon from '../../assets/icons/1046784.svg';
 
 const RestaurantTitle = () => {
   return (
